+++ src/setuptools_dso/test/test_probe.py
@@ -1,4 +1,5 @@
 
+import os
 import unittest
 
 from .. import probe
@@ -17,3 +18,45 @@ class TryCompile(unittest.TestCase):
         self.assertTrue(self.probe.check_symbol('RAND_MAX', headers=['stdlib.h']))
         self.assertTrue(self.probe.check_symbol('abort', headers=['stdlib.h']))
         self.assertFalse(self.probe.check_symbol('intentionally_undeclared_symbol', headers=['stdlib.h']))
+
+    def test_macros(self):
+        inp = os.path.join(self.probe.tempdir, 'defs.h')
+        with open(inp, 'w') as F:
+            F.write('''
+/* not defined UNDEF */
+#define NOVAL
+#define MAGIC 42
+#define HELLO "hello world"
+#define MULTILINE this \
+is a test
+''')
+
+        defs = self.probe.eval_macros(['UNDEF', 'NOVAL', 'MAGIC', 'HELLO', 'MULTILINE'], headers=[inp])
+
+        # GCC/clang == ' ', msvc == ''
+        if defs['NOVAL']=='':
+            defs['NOVAL'] = ' '
+
+        self.assertListEqual(list(defs.items()), [
+            ('UNDEF', None),
+            ('NOVAL', ' '),
+            ('MAGIC', '42'),
+            ('HELLO', '"hello world"'),
+            ('MULTILINE', 'this is a test'),
+        ])
+
+    def test_predef(self):
+        gnuc, clang, msc_ver = self.probe.eval_macros(['__GNUC__', '__clang__', '_MSC_VER'])
+        self.assertTrue(gnuc or clang or msc_ver)
+
+    def test_info(self):
+        info = self.probe.info
+        print("Raw Macros", info._raw_macros)
+        print("Info", info)
+
+        self.assertIn(info.compiler, ('clang', 'gcc', 'msvc'))
+        self.assertGreater(info.compiler_version, (0,))
+        self.assertIn(info.target_os, ("cygwin", "linux", "osx", "windows"))
+        self.assertIn(info.target_arch, ("aarch64", "arm32", "amd64", "i386"))
+        self.assertIn(info.address_width, (32, 64))
+        self.assertIn(info.endian, ("little", "big"))